More specs for invalidly encoded input
This time instead of directly testing the encoding transformation
functions we call the "parse" method and make sure that the expected
exception bubbles up.

This is in a sense a duplication of the existing tests but it is
important to confirm that the parse method behaves as expected in the
face of invalid input, so the double-check here really doesn't hurt.

Signed-off-by: Wincent Colaiuta <win@wincent.com>