Add spec to illustrate em/strong disambiguation
On seeing "'''''" the scanner has to decide whether it is seeing
"<strong><em>" or "<em><strong>". In order to be deterministic without
looking ahead it always chooses the former.

This commit adds a spec that shows how to use whitespace to disambiguate
in the case where you really do want to indicate "<em><strong>".

Signed-off-by: Wincent Colaiuta <win@wincent.com>