Revert ESCAPED_TT tokenization
I don't want to introduce special escapes into the grammar, even though
this one works. In the interests of consistency the correct solution is
to just embed literal backticks in a <nowiki></nowiki> span, as
demonstrated in the specs. Literal backticks are already passed through
unaltered in <pre></pre> and <tt>spans</tt> anyway, which is arguably
the most common case.

Signed-off-by: Wincent Colaiuta <win@wincent.com>