Replace ANTLR lexer with Ragel scanner
This insertion involves a rather massive refactoring so I'm doing it in
several steps: this first step is basically the result of ripping out
the old lexer and implanting the new scanner, up to the point were we
get an error and warning-free build.

From here on it's just a case of making adjustments to get all the specs
passing again.

Signed-off-by: Wincent Colaiuta <win@wincent.com>