Fix latent bug in CRLF case
Not sure why this bug never surfaced before with the ANTLR lexer, but
there was a latent bug in the CRLF case which was causing the paragraph
specs to fail. In the cases where the line array was empty consecutive
newlines weren't being interpreted as paragraph breaks.

I suspect the cause is actually in the setup prior to the switch
statement (probably not pusing the current token onto the line stack).

Signed-off-by: Wincent Colaiuta <win@wincent.com>