Fix UL spec breakage
This was an incredibly subtle bug that cost me a couple of hours to
analyze. Looking at the old (ANLTR) code, I can't actually figure out
how it produced correct results; it seems impossible.

The new code was basically identical to the old code but produced
different results. In trying to figure out why I analyzed what the old
code was doing and even now after a couple of hours on the problem I
can't see how the old code possibly worked, but mysteriously it did.

In any case now, I've implemented the logical fix for the new code and
it now works too. A nice side effect of this is that we automatically
gobble up more excess whitespace now; this isn't due to any change in
the parser, but in the lexer instead, because the later now returns
runs of spaces as a single token when appropriate instead of yielding
individual spaces.

Signed-off-by: Wincent Colaiuta <win@wincent.com>