Initial activation of indenting code
At the moment this breaks some of the specs but I wanted to get it in
the repository in any case as it looks to be very nearly complete.

Will follow this up with edits to the specs to make them pass, as
there are now many expected results that will need to be modified to
include the indentation.

Signed-off-by: Wincent Colaiuta <win@wincent.com>