Add indentation specs
Test user overrides of base indentation. The intention here is for
incorporation in Haml templates (or any other templating system that
takes care to produce beautifully-formatted HTML). The user can supply
a base indentation level that matches the expected indentation level
in the enclosing Haml template; in this way the wikitext-generated
portion won't break the nice indentation of the template as a whole.

Signed-off-by: Wincent Colaiuta <win@wincent.com>