Documentation: remove source code comments covered elsewhere
Remove redundant RDoc comments from the main wikitext module source file
as these are now covered in the separate RDoc file in the doc
subdirectory.

At the same time update the definition of "treat_slashes_as_special" to
reflect the new, stricter definition of what classifies as a special
link.

Signed-off-by: Wincent Colaiuta <win@wincent.com>