Bump version number post-release
The released version was 0.1. As far as I can tell this is basically a
feature-complete release (in terms of the features that I have
envisioned for 1.0), so any subsequent releases will be bug-fix-only
ones, probably following a numbering scheme like 0.2, 0.3 and so on.

The plan is to use the bundle in production, add more specs, correct
any issues that might come up, and go to 1.0 once a reasonable interval
has passed.

Signed-off-by: Wincent Colaiuta <win@wincent.com>