Fix non-indentation inside PRE blocks
If you try to indent material inside a PRE block you'll find that it
doesn't work. What's actually happening is that the inital spaces are
getting emitted on the previous line, _before_ the CRLF. The problem
therefore shows up as missing indentation, but the reality is that the
spaces _are_ being emitted, only in the wrong place.

This commit applies the minimal fix for the issue, but there is more to
come. I noticed the issue with spaces but it could happen with any
token. The only reason I hadn't noticed up until now is that the most
common case (the PRINTABLE token) is correctly handled because that rule
calls the "start_para_if_necessary" function, and that emits any pending
CRLFs that might be needed.

The next commit will generalize this fix to other token types and add
specs for them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>