Fix all known pending CRLF emission bugs
This is the generalization of the fix in 191b75d (which specifically
applied to the SPACE token). We apply the same fix to all other token
types where it is required, with specs to verify that the fix actually
works. We also add "future-proofing" specs to those token types where
the fix is _not_ required, to be sure of catching the issue if it
every crops up.

Signed-off-by: Wincent Colaiuta <win@wincent.com>