Handle auto-linking in when followed by "special" URI characters
This is something which (I recall) worked under the ANTLR implementation
but which didn't survive in the move to Ragel. Basically, the greedy
match behaviour of the Ragel scanner ensured that whenever an email
address or URI was followed by a "special" URI character (that is a
character which is valid when _within_ a URI but which we actually want
to server as a delimiter when it appears as the last thing in a URI)
then it would actually scan the entire thing as a PRINTABLE token rather
than a URI followed by a PRINTABLE.

For example:

  I think http://example.com/ is a great website

Would work because the SPACE tokens on either side of the URI would
delimit it and the scanner would choose URI rather than PRINTABLE
because URI is declared first in the grammar.

In contrast, an input string like this:

  The best one is http://example.com/.

Would tokenize this as PRINTABLE, SPACE, PRINTABLE, SPACE, PRINTABLE,
SPACE, PRINTABLE, SPACE, PRINTABLE, because "http://example.com/."
matches the PRINTABLE machine and is a longer match than
"http://example.com/" in the URI machine.

The solution is to chop out the "special URI chars" from the PRINTABLE
class. This requires minimal changes to the grammar, and at the parser
level the most difficult part of the change is updating the processing
of image syntax (because that previously acceptable only a single
PRINTABLE token and now it must scan ahead for both PRINTABLE and
SPECIAL_URI_CHARS tokens in a loop).

These changes add a (small) amount of complexity, but before-and-after
benchmarking shows that the impact is insignificant.

I've added a bunch of specs which reveal the old problem and will help
us to catch it if it every crops up again.

Signed-off-by: Wincent Colaiuta <win@wincent.com>